+++ src/bedrock/world/actor/player/player.h
@@ -196,7 +196,7 @@ public:
 
     // Endstone begins
     BedSleepingResult getBedResult(BlockPos const &bed_pos);  // moved from startSleepInBed into separate method
-    [[nodiscard]] const std::shared_ptr<ContainerManagerModel> &getContainerManagerModel() const
+    [[nodiscard]] const std::shared_ptr<ContainerManagerModel> &getContainerManagerModelRef() const
     {
         return container_manager_;
     }