+++ src/bedrock/world/inventory/network/crafting/item_stack_request_action_craft_handler.h
@@ -36,7 +36,8 @@ public:
 
     std::vector<std::pair<ItemInstance, unsigned int>> &getCraftResults() { return craft_results_; }  // Endstone
 
-    [[nodiscard]] ItemStackRequestActionHandler &getRequestActionHandler() const { return request_action_handler_; }
+    [[nodiscard]] const ItemStackRequestActionHandler &getActionHandler() const { return request_action_handler_; }
+    [[nodiscard]] ItemStackRequestActionHandler &getActionHandler() { return request_action_handler_; }  // Endstone
 
 private:
     ItemStackRequestActionHandler &request_action_handler_;