+++ tests/plugin/src/endstone_test/listeners/inventory_event_listener.py
@@ -1,4 +1,9 @@
-from endstone.event import InventoryCloseEvent, InventoryOpenEvent, event_handler
+from endstone.event import (
+    EnchantItemEvent,
+    InventoryCloseEvent,
+    InventoryOpenEvent,
+    event_handler,
+)
 
 from .event_listener import EventListener
 
@@ -20,4 +25,22 @@ class InventoryEventListener(EventListener):
             f"{event.player.name} closed an inventory of {event.inventory.size} slots",
             player=event.player.name,
             size=event.inventory.size,
+        )
+
+    @event_handler
+    def on_enchant_item(self, event: EnchantItemEvent):
+        self.record(
+            event,
+            f"{event.enchanter.name} enchants {event.item} using option "
+            f"{event.which_button}",
+            always_log=True,
+            player=event.enchanter.name,
+            block_type=str(event.enchant_block.type),
+            item_type=str(event.item.type),
+            exp_level_cost=event.exp_level_cost,
+            enchants={
+                str(enchantment.id): level
+                for enchantment, level in event.enchants_to_add.items()
+            },
+            which_button=event.which_button,
         )